requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-gitlab==3.15.0
jira==3.5.2
langchain==0.0.335
boto3==1.28.62
celery==5.3.4
redis==5.0.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
pylint==3.0.2
eslint==0.0.1
flake8==6.1.0
mypy==1.6.1
black==23.10.0
isort==5.12.0
pyyaml==6.0.1
kubernetes==28.1.0